requirements.txt
mysql-connector-python==8.0.33
neo4j==5.9.0
requests==2.31.0
pandas==2.0.2